+++ pkg/gofr/external_db.go
@@ -14,6 +14,15 @@ func (a *App) AddMongo(db container.MongoProvider) {
 	a.container.Mongo = db
 }
 
+func (a *App) AddFTP(fs container.FileSystemProvider) {
+	fs.UseLogger(a.Logger())
+	fs.UseMetrics(a.Metrics())
+
+	fs.Connect()
+
+	a.container.File = fs
+}
+
 // AddClickhouse initializes the clickhouse client.
 // Official implementation is available in the package : gofr.dev/pkg/gofr/datasource/clickhouse .
 func (a *App) AddClickhouse(db container.ClickhouseProvider) {
