--- pkg/gofr/file_interface/go.mod
@@ -1,3 +1,0 @@
-module gofr.dev/pkg/gofr/interface
-
-go 1.22.4
